Updated for 1.1.40428.1602, added services
Put much more meat on the bones of this now.  The original client class
has been effectively moved down a level of abstraction to being merely
an HTTP workhorse (the original class name is now obsolete btw, but the
compiler error you'll get explains what to do); with a new IClient
abstraction + implementation, plus a load of more 'structured' services
to provide access to the static data used by the services over cached
copies retrieved once for the lifetime of each service.

Added loads more unit tests too - which also serve to demonstrate how to
use each of the services.

There might be some documentation holes - I apologise for that - I
finished this in a hurry as I had to move on to something else.